Extract UDP response-rate-limiting policy out of DnsServer
